Fixed some errors in Simulation and watered checking.
Fixed a bug where pipes were not being placed or cloned.

Fixed a bug where Sustainability was not being calculated correctly for
plants.

Fixed some bugs with how we check if a plant is Watered.

This appears to still have some remaining bug(s) - we sometimes see a
plant getting watered, sometimes not, when we would expect it to be.