Simulation now updates current state, rather than replacing.
Simulation creates a copy of the state, but now it uses it
as an "old state" and modifies the current state.

Previously, it would copy the current state and then replace
the existing state with the copy. This caused references in
UI and visualization code to be pointing to the old state,
and never updated.

While it would be better to never hold onto the current state,
and instead ask for it when we need it, that will take a lot
more code changes. So this should work for now.